Fix test file calling close test
Obviously the Correct(tm) fix is to just start using
tox or another test runner which can automatically
run pytest with multiple Python versions, instead of
the current hack which relies on invoking the test
file as a script and thus costs __main__ maintenance.

But I just don't have the time or the drive to get
into that right now.