Rename yield_from to YieldFrom
The original reason to not do this was conceptualizing
the interface as more of a function that returned some
magic wrapper iterator.

However, usage has shown that the interface really is
a class, not just incidentally implemented as a class:
more complex use-cases end up needing to treat it as
a class, and the temptation to subclass it to fulfill
the same interface easily has come up at least once.