rejected idea scraps/stash/junk
The prior commits in this branch were building up to a v2 release
and might be worth pursuing again when I have more time, although
it is worth rethinking them now that I've more fully understood
exactly how ``yield from`` works in CPython and iin various other
Pythons.

This commit is just scraps and drafts from an earlier detour into
implementing full generator emulation and implementing `iwith`
(see `with-as-a-function` on PyPI or `python-with` on my GitHub),
which I had initially thought I would implement on top of a more
general base which I initially thought I would put into this
package. By now I think I will eventually do it in a separate
package, but I don't have the time for that right now.